requirements.txt
nonebot2[fastapi]
nonebot-adapter-onebot
elasticsearch
aiopytesseract
aiofiles
httpx
pillow
torch
torchvision
transformers